Created Image upload/download logic
The Image up- and download logic has been created and tested. For this to work a service FileService has been created. The images are stored on the server (in this case my laptop) and the path to the image is stored in the MediaFile db model. This service is currently used by the MarketplaceService and ShopService.